Massive Restructuring update
 - Split kvs_webrtc into multiple components: app_webrtc, kvs_signaling and kvs_webrtc for kvs_peer
 - app_webrtc now takes peer_connection_if and signaling_if both as config args
 - Updated the webrtc_classic example with these restructuring changes

 - TODO: Some minor dependencies on KVS SDK sources remain

+++ esp_port/examples/signaling_only/main/signaling_only_main.c
@@ -201,7 +201,7 @@ void app_main(void)
 
     status = app_webrtc_init(&app_webrtc_config);
     if (status != WEBRTC_STATUS_SUCCESS) {
-        ESP_LOGE(TAG, "Failed to initialize WebRTC application: 0x%08" PRIx32, status);
+        ESP_LOGE(TAG, "Failed to initialize WebRTC application: 0x%08" PRIx32, (uint32_t) status);
         return;
     }
 
@@ -210,7 +210,7 @@ void app_main(void)
 
     status = app_webrtc_run();
     if (status != WEBRTC_STATUS_SUCCESS) {
-        ESP_LOGE(TAG, "WebRTC application failed: 0x%08" PRIx32, status);
+        ESP_LOGE(TAG, "WebRTC application failed: 0x%08" PRIx32, (uint32_t) status);
         app_webrtc_terminate();
     } else {
         ESP_LOGI(TAG, "WebRTC application started successfully");
